app/src/main/java/com/xunobulax/rambutan/data/Person.kt
package com.xunobulax.rambutan.data

import androidx.room.ColumnInfo
import androidx.room.Entity
import androidx.room.PrimaryKey
import java.time.LocalDate


@Entity(tableName = "people")
data class Person(
    @PrimaryKey(autoGenerate = true)
    val id: Int,

    @ColumnInfo(name = "first_name")
    var firstName: String?,

    @ColumnInfo(name = "last_name")
    var lastName: String?,

    var email: String?,

    var birthday: LocalDate?,

    @ColumnInfo(name = "partner_id")
    var partnerId: Int?
)